refactor: 파일명 및 모듈 경로 수정
서비스 및 컨트롤러 관련 파일명을 더 직관적으로 변경하고, 관련 모듈 경로를 통일성 있게 정리했습니다. 이를 통해 코드 가독성을 향상시키고 유지보수를 용이하게 만들었습니다.
